+++ app/helpers/artificial-intelligence.js
@@ -1,49 +1,37 @@
 let utils = require('../helpers/utils');
 
-function setup(state, sprite) {
+function setup(state, $sprite) {
     function walkingLeft() {
         utils.timesRandomAsync(state, 10, Phaser.Timer.SECOND / 20, () => {
-            sprite.events.onLeft.dispatch();
+            $sprite.events.onLeft.dispatch();
         });
     }
 
     function walkingRight() {
         utils.timesRandomAsync(state, 10, Phaser.Timer.SECOND / 20, () => {
-            sprite.events.onRight.dispatch();
+            $sprite.events.onRight.dispatch();
         });
     }
 
     function boxing() {
         utils.timesRandomAsync(state, 5, Phaser.Timer.SECOND / 4, () => {
-            sprite.events.onBoxing.dispatch();
+            $sprite.events.onBoxing.dispatch();
         });
     }
 
     function kicking() {
         utils.timesRandomAsync(state, 5, Phaser.Timer.SECOND / 4, () => {
-            sprite.events.onKicking.dispatch();
+            $sprite.events.onKicking.dispatch();
         });
     }
 
     function jumping() {
-        sprite.events.onJumping.dispatch();
+        $sprite.events.onJumping.dispatch();
     }
-
-    let moves = [
-        walkingLeft, walkingRight
-    ];
-
-    let dodge = [
-        jumping
-    ];
-
-    let fight = [
-        boxing, kicking
-    ];
 
     function start(intervalTime, strategy) {
         utils.interval(state, intervalTime, () => {
-            if (!sprite.alive) {
+            if (!$sprite.alive) {
                 return;
             }
 
@@ -54,9 +42,9 @@ function setup(state, sprite) {
         });
     }
 
-    start(Phaser.Timer.SECOND, fight);
-    start(Phaser.Timer.SECOND / 2, moves);
-    start(Phaser.Timer.SECOND / 3, dodge);
+    start(Phaser.Timer.SECOND, [boxing, kicking]);
+    start(Phaser.Timer.SECOND / 3, [walkingLeft, walkingRight]);
+    start(Phaser.Timer.SECOND * 5, [jumping]);
 }
 
 module.exports = {