+++ app/states/loading-state.js
@@ -34,19 +34,19 @@ export default class LoadingState extends Phaser.State {
         this.load.image('bg-credits', 'backgrounds/credits/credits.png');
         this.load.image('bg-select-player', 'backgrounds/select-player/select-player.png');
 
-        this.load.image('son-goku-card', 'posters/son-goku/son-goku-card.png');
-        this.load.image('vegeta-card', 'posters/vegeta/vegeta-card.png');
-        this.load.image('bubu-card', 'posters/bubu/bubu-card.png');
-        this.load.image('cell-card', 'posters/cell/cell-card.png');
-        this.load.image('freeza-card', 'posters/freeza/freeza-card.png');
-        this.load.image('piccolo-card', 'posters/piccolo/piccolo-card.png');
+        this.load.image('son-goku-card', 'posters/son-goku/son-goku-card-160x240.png');
+        this.load.image('vegeta-card', 'posters/vegeta/vegeta-card-160x240.png');
+        this.load.image('bubu-card', 'posters/bubu/bubu-card-160x240.png');
+        this.load.image('cell-card', 'posters/cell/cell-card-160x240.png');
+        this.load.image('freeza-card', 'posters/freeza/freeza-card-160x240.png');
+        this.load.image('piccolo-card', 'posters/piccolo/piccolo-card-160x240.png');
 
         this.load.spritesheet('son-goku-fight', 'fight/son-goku/son-goku-spritesheet.png', 150, 200);
-        this.load.spritesheet('vegeta-fight', 'fight/vegeta/vegeta-fight.png', 150, 200);
-        this.load.spritesheet('bubu-fight', 'fight/bubu/bubu-fight.png', 150, 200);
-        this.load.spritesheet('cell-fight', 'fight/cell/cell-fight.png', 150, 200);
-        this.load.spritesheet('freeza-fight', 'fight/freeza/freeza-fight.png', 150, 200);
-        this.load.spritesheet('piccolo-fight', 'fight/piccolo/piccolo-fight.png', 150, 200);
+        this.load.spritesheet('vegeta-fight', 'fight/vegeta/vegeta-spritesheet.png', 150, 200);
+        this.load.spritesheet('bubu-fight', 'fight/bubu/bubu-spritesheet.png', 150, 200);
+        this.load.spritesheet('cell-fight', 'fight/cell/cell-spritesheet.png', 150, 200);
+        this.load.spritesheet('freeza-fight', 'fight/freeza/freeza-spritesheet.png', 150, 200);
+        this.load.spritesheet('piccolo-fight', 'fight/piccolo/piccolo-spritesheet.png', 150, 200);
 
         this.load.spritesheet('collect-spritesheet', 'sheets/collect-spritesheet.png', 40, 40);
         this.load.spritesheet('ball-spritesheet', 'sheets/ball-spritesheet.png', 40, 40);